+++ package.json
@@ -14,16 +14,18 @@
     "cssrecipes-vertical-rhythm": "^0.6.0",
     "invariant": "^2.1.1",
     "normalize.css": "^3.0.3",
-    "react": "^0.14.0",
+    "react": "^0.14.7",
     "react-disqus-thread": "^0.3.1",
-    "react-dom": "^0.14.0",
+    "react-dom": "^0.14.7",
     "react-google-analytics": "^0.2.0",
-    "react-helmet": "^2.1.1",
-    "react-redux": "^2.1.2",
-    "react-router": "^1.0.0-rc3",
-    "redux": "^2.0.0",
-    "redux-devtools": "^2.1.5",
-    "redux-thunk": "^0.1.0",
+    "react-helmet": "^2.2.0",
+    "react-redux": "^4.0.0",
+    "react-router": "^2.0.0",
+    "redux": "^3.0.0",
+    "redux-devtools": "^3.0.1",
+    "redux-devtools-dock-monitor": "^1.0.1",
+    "redux-devtools-log-monitor": "^1.0.1",
+    "redux-thunk": "^1.0.3",
     "whatwg-fetch": "^0.9.0"
   },
   "devDependencies": {
@@ -31,6 +33,7 @@
     "babel-core": "^6.5.2",
     "babel-eslint": "^5.0.0",
     "babel-loader": "^6.2.3",
+    "babel-plugin-webpack-loaders": "^0.3.6",
     "babel-polyfill": "^6.5.0",
     "babel-preset-es2015": "^6.5.0",
     "babel-preset-react": "^6.5.0",
@@ -38,6 +41,7 @@
     "babel-preset-stage-1": "^6.5.0",
     "chalk": "^1.0.0",
     "classnames": "^1.2.0",
+    "cross-env": "^1.0.7",
     "css-loader": "^0.17.0",
     "eslint": "^1.10.0",
     "eslint-loader": "^1.1.0",
@@ -62,7 +66,7 @@
     "promise": "^7.0.0",
     "raw-loader": "^0.5.1",
     "rimraf": "^2.3.2",
-    "statinamic": "^0.6.0",
+    "statinamic": "^0.7.0",
     "style-loader": "^0.12.2",
     "svgo": "^0.6.0",
     "svgo-loader": "^1.1.0",
@@ -74,14 +78,14 @@
   },
   "scripts": {
     "lint": "eslint .",
-    "start": "babel-node scripts/putaindebuild --server --open --dev",
-    "build": "babel-node scripts/putaindebuild --static --production",
-    "postbuild": "echo 'putaindecode.io' > dist/CNAME && touch dist/.nojekyll",
+    "statinamic": "cross-env BABEL_ENV=statinamic DEBUG=statinamic:* babel-node scripts/build",
+    "start": "npm run statinamic -- --server --open --dev",
+    "build": "npm run statinamic -- --static --production",
     "pretest": "npm run lint",
     "test": "npm run build"
   },
-  "browser": {
-    "fs": false
+  "statinamic": {
+    "CNAME": true
   },
   "babel": {
     "presets": [
@@ -90,9 +94,15 @@
       "babel-preset-stage-1"
     ],
     "env": {
-      "development": {
-        "presets": [
-          "babel-preset-react-hmre"
+      "statinamic": {
+        "plugins": [
+          [
+            "babel-plugin-webpack-loaders",
+            {
+              "config": "./scripts/webpack.config.babel.js",
+              "verbose": false
+            }
+          ]
         ]
       }
     }
@@ -159,12 +169,6 @@
       "react/self-closing-comp": 2,
       "react/sort-comp": 0,
       "react/wrap-multilines": 2
-    },
-    "globals": {
-      "__BASE_URL__": true,
-      "__DEV__": true,
-      "__PROD__": true,
-      "__DEVTOOLS__": true
     },
     "plugins": [
       "react"
