+++ web_modules/GoogleAnalyticsTracker/index.js
@@ -4,16 +4,31 @@ import { PropTypes } from "react"
 import ga from "react-google-analytics"
 const GoogleAnalyticsInitiailizer = ga.Initializer
 
-class GoogleAnalyticsTracker extends Component {
+const isProduction = process.env.NODE_ENV === "production"
+const isClient = typeof window !== "undefined"
+
+export default class GoogleAnalyticsTracker extends Component {
+
+  static propTypes = {
+    children: PropTypes.oneOfType([ PropTypes.array, PropTypes.object ]),
+    params: PropTypes.object.isRequired,
+  };
+
+  static contextTypes = {
+    metadata: PropTypes.object.isRequired,
+  };
+
   componentWillMount() {
-    const { pkg } = this.context.metadata
-    if (__PROD__) {
-      ga("create", pkg.googleAnalyticsUA, "auto")
+    if (isClient) {
+      const { pkg } = this.context.metadata
+      if (isProduction) {
+        ga("create", pkg.googleAnalyticsUA, "auto")
+      }
+      else {
+        console.info("ga.create", pkg.googleAnalyticsUA)
+      }
+      this.logPageview()
     }
-    if (__DEV__) {
-      console.info("ga.create", pkg.googleAnalyticsUA)
-    }
-    this.logPageview()
   }
 
   componentWillReceiveProps(props) {
@@ -23,11 +38,13 @@ class GoogleAnalyticsTracker extends Component {
   }
 
   logPageview() {
-    if (__PROD__) {
-      ga("send", "pageview")
-    }
-    if (__DEV__ && typeof window !== "undefined") {
-      console.info("New pageview", window.location.href)
+    if (isClient) {
+      if (isProduction) {
+        ga("send", "pageview")
+      }
+      else {
+        console.info("New pageview", window.location.href)
+      }
     }
   }
 
@@ -40,14 +57,3 @@ class GoogleAnalyticsTracker extends Component {
     )
   }
 }
-
-GoogleAnalyticsTracker.propTypes = {
-  children: PropTypes.oneOfType([ PropTypes.array, PropTypes.object ]),
-  params: PropTypes.object.isRequired,
-}
-
-GoogleAnalyticsTracker.contextTypes = {
-  metadata: PropTypes.object.isRequired,
-}
-
-export default GoogleAnalyticsTracker
