+++ web_modules/Layouts/index.js
@@ -1,0 +1,6 @@
+export Page from "./Page"
+export PageError from "./PageError"
+export Homepage from "./Homepage"
+export Post from "./Post"
+export Posts from "./Posts"
+export AdventCalendar from "./AdventCalendar"
