--- web_modules/app/pageComponents.js
@@ -1,6 +1,0 @@
-export Page from "Layout/Page"
-export PageError from "Layout/PageError"
-export Homepage from "Layout/Homepage"
-export Post from "Layout/Post"
-export Posts from "Layout/Posts"
-export AdventCalendar from "Layout/AdventCalendar"
