+++ web_modules/app/routes.js
@@ -2,12 +2,12 @@ import React from "react"
 import { Route } from "react-router"
 import PageContainer from "statinamic/lib/PageContainer"
 
-// components
-import Layout from "Layout"
+// Layouts
+import LayoutContainer from "../LayoutContainer"
 
 // routes
 export default (
-  <Route component={ Layout }>
+  <Route component={ LayoutContainer }>
     <Route path="*" component={ PageContainer } />
   </Route>
 )
