+++ web_modules/i18n/index.js
@@ -1,4 +1,4 @@
 export default {
-  en: require("json-loader!yaml-loader!./en.yml"),
-  fr: require("json-loader!yaml-loader!./fr.yml"),
+  en: require("./en.yml"),
+  fr: require("./fr.yml"),
 }
